Renamed error module to errors
